requirements.txt
nlp-pie==0.2.2
flask
pyyaml
regex==2019.8.19